refactor: config 분리
