+++ src/main/kotlin/io/concert/booking/BookingApplication.kt
@@ -6,8 +6,6 @@ import org.springframework.boot.runApplication
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 import org.springframework.scheduling.annotation.EnableScheduling
 
-@EnableJpaAuditing
-@EnableScheduling
 @EnableConfigurationProperties
 @SpringBootApplication
 class BookingApplication
