+++ src/main/kotlin/io/concert/booking/config/jpa/JpaConfig.kt
@@ -1,0 +1,9 @@
+package io.concert.booking.config.jpa
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
+
+@Configuration
+@EnableJpaAuditing
+class JpaConfig {
+}
