+++ src/main/kotlin/io/concert/booking/config/scheduler/SchedulerConfig.kt
@@ -1,0 +1,9 @@
+package io.concert.booking.config.scheduler
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.scheduling.annotation.EnableScheduling
+
+@Configuration
+@EnableScheduling
+class SchedulerConfig {
+}
