+++ src/test/kotlin/io/concert/booking/infra/queue/TokenSchedulerTest.kt
@@ -1,6 +1,6 @@
 package io.concert.booking.infra.queue
 
-import io.concert.booking.config.properties.SchedulerProperties
+import io.concert.booking.config.scheduler.SchedulerProperties
 import io.concert.booking.domain.queue.FakeTokenRepository
 import io.concert.booking.domain.queue.Token
 import io.concert.booking.domain.queue.TokenStatus
